+++ django_fsm/apps.py
@@ -1,0 +1,8 @@
+from __future__ import annotations
+
+from django.apps import AppConfig
+
+
+class DjangoFSMAppConfig(AppConfig):
+    name = "django_fsm"
+    verbose_name = "Django FSM"